mm: process_reclaim: Rectify the LOWMEM_CRITICAL condition
Before, we took the sum of ACTIVE and INACTIVE file-pages and based on
whether that sum was greater than free_file_limit, we chose to trigger
the LOWMEM_CRITICAL condition.

The above approach proves to be slow in the case where swap suddenly
depletes and goes to 0,and active+inactive filepages is still under the
free_file_limit.The result is that major system freezes occur,
temporarily, for a period of time.

In order to react faster, we can consider only the ACTIVE file-pages,
and modify the triggering condition such that when the number of
ACTIVE file-pages are below free_file_limit,this will trigger the
LOWMEM_CRITICAL scenario.This fixes the problem as stated above.

(Thanks to @markvgsilva for debugging the problem extensively)