selinux: Allow enforcing mode spoofing
This is a developer-only feature. Apply only if your SEPolicy is borked.

Some ROMS are released with bad or unfixed sepolicies, and fixing them
can sometimes be a huge headache to the developer.So, for their peace of
mind, albeit temporarily, we can spoof enforcing so that denials will
only be logged and not enforced, and at the same time, SafetyNet and
SafetyNet-dependent applications will continue to work.

To spoof enforcing, just run:
echo 2 > /sys/fs/selinux/enforce

Or if you want to spoof enforcing on boot, add this to the kernel
cmdline:
enforcing=2

P.S: This patch has saved many a developer from jumping off a building,
after getting extremely livid due to stupid sepolicies.